+++ src/Edde/Api/Schema/ILink.php
@@ -4,14 +4,6 @@
 
 		interface ILink {
 			/**
-			 * schema of a link contains name of a link and could
-			 * contain link properties
-			 *
-			 * @return ISchema
-			 */
-			public function getSchema(): ISchema;
-
-			/**
 			 * get a from source schema/property
 			 *
 			 * @return ITarget
