+++ src/Edde/Api/Schema/IRelation.php
@@ -1,0 +1,11 @@
+<?php
+	namespace Edde\Api\Schema;
+
+		interface IRelation extends ILink {
+			/**
+			 * return the relation schema
+			 *
+			 * @return ISchema
+			 */
+			public function getSchema(): ISchema;
+		}
