+++ src/Edde/Common/Schema/Link.php
@@ -3,15 +3,10 @@
 	namespace Edde\Common\Schema;
 
 		use Edde\Api\Schema\ILink;
-		use Edde\Api\Schema\ISchema;
 		use Edde\Api\Schema\ITarget;
 		use Edde\Common\Object\Object;
 
 		class Link extends Object implements ILink {
-			/**
-			 * @var ISchema
-			 */
-			protected $schema;
 			/**
 			 * @var ITarget
 			 */
@@ -21,17 +16,9 @@
 			 */
 			protected $to;
 
-			public function __construct(ISchema $schema, ITarget $from, ITarget $to) {
-				$this->schema = $schema;
+			public function __construct(ITarget $from, ITarget $to) {
 				$this->from = $from;
 				$this->to = $to;
-			}
-
-			/**
-			 * @inheritdoc
-			 */
-			public function getSchema(): ISchema {
-				return $this->schema;
 			}
 
 			/**
